REF: Create test_encoding file for CSV
